moved class MailConfig into mail submodule Also fixed other files so that now the script runs completely
 Changes to be committed:
   (use "git reset HEAD <file>..." to unstage)

	modified:   certmon.py
	new file:   mail/mail_config.py
	modified:   mail/mail_list.py
	modified:   mail/mailer.py
